requirements.txt
aiofiles==0.5.0
bcrypt==3.2.0
fastapi==0.63.0
msal==1.18.0
oauthlib==3.2.0
pandas==1.4.2
passlib==1.7.4
psycopg2==2.9.3
python-dotenv==0.15.0
python-multipart==0.0.5
SQLAlchemy==1.3.22
uvicorn==0.13.3